requirements.txt
pandas
numpy
tensorflow
scikit-learn
streamlit
streamlit-drawable-canvas
Pillow